fix(ffi): close 4 murk-ffi bugs (#81, #78, #80, #66) — dedup, ownership, ABI, Sync
- #81: Extract shared `convert_obs_entry` in obs.rs; batched.rs imports it
- #78: Consume config handle before null-check in `murk_lockstep_create`
- #80: Replace `usize` with `u64` in repr(C) structs; add compile-time
  size/alignment assertions for MurkStepMetrics, MurkStepContext,
  MurkCommand, MurkReceipt, MurkWriteDecl, MurkObsEntry, MurkObsResult
- #66: Document deliberate `!Sync` on CallbackPropagator; identify
  `Mutex<LockstepWorld>` as the load-bearing serialization invariant

+++ crates/murk-python/src/metrics.rs
@@ -95,7 +95,7 @@ impl StepMetrics {
             total_us: m.total_us,
             command_processing_us: m.command_processing_us,
             snapshot_publish_us: m.snapshot_publish_us,
-            memory_bytes: m.memory_bytes,
+            memory_bytes: m.memory_bytes as usize,
             propagator_us,
         }
     }
